Matematicas

+++ Matematicas/BigNumbers.cpp
@@ -8,10 +8,10 @@
 #include <utility>
 using namespace std;
 
-typedef string BigInt;
+typedef string BigInteger;
 
 //Regresa el i-esimo digito de derecha a izquierda de un numero.
-unsigned int digit(const BigInt &num, unsigned int i) {
+unsigned int digit(const BigInteger &num, unsigned int i) {
     if (i < num.size())
         return num[num.size() - 1 - i] - '0';
     return 0;
@@ -19,7 +19,7 @@ unsigned int digit(const BigInt &num, unsigned int i) {
 
 //Compara dos numeros y regresa: 1 si el primero es mayor; 0 si son iguales; -1 si el 
 //segundo es mayor.
-int compareTo(const BigInt &a, const BigInt &b) {
+int compareTo(const BigInteger &a, const BigInteger &b) {
     for (int i = max(a.size(), b.size()) - 1; i >= 0; --i) {
         if (digit(a, i) > digit(b, i))
             return 1;
@@ -30,8 +30,8 @@ int compareTo(const BigInt &a, const BigInt &b) {
 }
 
 //Regresa la suma de dos numeros.
-BigInt sum(const BigInt &a, const BigInt &b) {
-    BigInt ans;
+BigInteger sum(const BigInteger &a, const BigInteger &b) {
+    BigInteger ans;
     int carry = 0;
     for (int i = 0; i < max(a.size(), b.size()); ++i) {
         carry += digit(a, i) + digit(b, i);
@@ -45,8 +45,8 @@ BigInt sum(const BigInt &a, const BigInt &b) {
 }
 
 //Regresa la diferencia de dos numeros. El primero debe ser mayor o igual que el segundo.
-BigInt substract(const BigInt &a, const BigInt &b) {
-    BigInt ans;
+BigInteger substract(const BigInteger &a, const BigInteger &b) {
+    BigInteger ans;
     int carry = 0;
     for (int i = 0; i < a.size(); ++i) {
         carry += digit(a, i) - digit(b, i);
@@ -65,11 +65,11 @@ BigInt substract(const BigInt &a, const BigInt &b) {
     return ans;
 }
 
-//Regresa el producto de dos numeros (BigInt x int).
-BigInt multiply(const BigInt &a, unsigned int b) {
+//Regresa el producto de dos numeros (BigInteger x int).
+BigInteger multiply(const BigInteger &a, unsigned int b) {
     if (b == 0)
         return "0";
-    BigInt ans;
+    BigInteger ans;
     int carry = 0;
     for (int i = 0; i < a.size(); ++i) {
         carry += digit(a, i) * b;
@@ -84,17 +84,17 @@ BigInt multiply(const BigInt &a, unsigned int b) {
     return ans;
 }
 
-//Regresa el producto de dos numeros (BigInt x BigInt).
-BigInt multiply(const BigInt &a, const BigInt &b) {
-    BigInt ans;
+//Regresa el producto de dos numeros (BigInteger x BigInteger).
+BigInteger multiply(const BigInteger &a, const BigInteger &b) {
+    BigInteger ans;
     for (int i = 0; i < b.size(); ++i)
         ans = sum(ans, multiply(a, digit(b, i)).append(i, '0'));
     return ans;
 }
 
-//Regresa el cociente y el residuo de la division (BigInt / int).
-pair<BigInt, unsigned int> divide(const BigInt &a, unsigned int b) {
-    pair<BigInt, int> ans;
+//Regresa el cociente y el residuo de la division (BigInteger / int).
+pair<BigInteger, unsigned int> divide(const BigInteger &a, unsigned int b) {
+    pair<BigInteger, int> ans;
     for (int i = a.size() - 1; i >= 0; --i) {
         ans.second = 10*ans.second + digit(a, i);
         if (!ans.first.empty() || ans.second >= b || i == 0)
@@ -104,10 +104,10 @@ pair<BigInt, unsigned int> divide(const BigInt &a, unsigned int b) {
     return ans;
 }
 
-//Regresa el cociente y el residuo de la division (BigInt / BigInt).
-pair<BigInt, BigInt> divide(const BigInt &a, const BigInt &b) {
-    pair<BigInt, BigInt> ans;
-    BigInt table[10];
+//Regresa el cociente y el residuo de la division (BigInteger / BigInteger).
+pair<BigInteger, BigInteger> divide(const BigInteger &a, const BigInteger &b) {
+    pair<BigInteger, BigInteger> ans;
+    BigInteger table[10];
     for (int i = 0; i < 10; ++i)
         table[i] = multiply(b, i);
     for (int i = a.size() - 1; i >= 0; --i) {
@@ -123,7 +123,7 @@ pair<BigInt, BigInt> divide(const BigInt &a, const BigInt &b) {
 }
 
 int main() {
-    BigInt a, b;
+    BigInteger a, b;
     cin >> a >> b;
 
     cout << a << " + " << b << " = " << sum(a, b) << '\n';
